[Update] Modify background task
