--- I2SN/I2SN/Delegates/BackgroundDelegate.swift
@@ -1,12 +1,0 @@
-//
-//  BackgroundDelegate.swift
-//  I2SN
-//
-//  Created by 권준원 on 2021/02/15.
-//
-
-import Foundation
-
-protocol BackgroundDelegate {
-    func notifyRemainTime()
-}
